+++ lib/config.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace Roots\Sage\Config;
+
 use Roots\Sage\Sidebar;
 
 /**
@@ -41,28 +42,28 @@ function display_sidebar() {
   static $display;
 
   if (!isset($display)) {
-    $sidebar_config = new Sidebar\Sage_Sidebar(
+    $sidebar_config = new Sidebar\SageSidebar(
       /**
        * Conditional tag checks (http://codex.wordpress.org/Conditional_Tags)
        * Any of these conditional tags that return true won't show the sidebar
        *
        * To use a function that accepts arguments, use the following format:
        *
-       * array('function_name', array('arg1', 'arg2'))
+       * ['function_name', ['arg1', 'arg2']]
        *
        * The second element must be an array even if there's only 1 argument.
        */
-      array(
+      [
         'is_404',
         'is_front_page'
-      ),
+      ],
       /**
        * Page template checks (via is_page_template())
        * Any of these page templates that return true won't show the sidebar
        */
-      array(
+      [
         'template-custom.php'
-      )
+      ]
     );
     $display = apply_filters('sage/display_sidebar', $sidebar_config->display);
   }