+++ lib/nav.php
@@ -9,31 +9,30 @@ namespace Roots\Sage\Nav;
  *   <li id="menu-item-8" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-8"><a href="/">Home</a></li>
  *   <li id="menu-item-9" class="menu-item menu-item-type-post_type menu-item-object-page menu-item-9"><a href="/sample-page/">Sample Page</a></l
  *
- * Sage_Nav_Walker example output:
+ * SageNavWalker example output:
  *   <li class="menu-home"><a href="/">Home</a></li>
  *   <li class="menu-sample-page"><a href="/sample-page/">Sample Page</a></li>
  */
-class Sage_Nav_Walker extends \Walker_Nav_Menu {
-  function check_current($classes) {
+class SageNavWalker extends \Walker_Nav_Menu {
+  public function checkCurrent($classes) {
     return preg_match('/(current[-_])|active|dropdown/', $classes);
   }
 
-  function start_lvl(&$output, $depth = 0, $args = array()) {
+  // @codingStandardsIgnoreStart
+  function start_lvl(&$output, $depth = 0, $args = []) {
     $output .= "\n<ul class=\"dropdown-menu\">\n";
   }
 
-  function start_el(&$output, $item, $depth = 0, $args = array(), $id = 0) {
+  function start_el(&$output, $item, $depth = 0, $args = [], $id = 0) {
     $item_html = '';
     parent::start_el($item_html, $item, $depth, $args);
 
     if ($item->is_dropdown && ($depth === 0)) {
       $item_html = str_replace('<a', '<a class="dropdown-toggle" data-toggle="dropdown" data-target="#"', $item_html);
       $item_html = str_replace('</a>', ' <b class="caret"></b></a>', $item_html);
-    }
-    elseif (stristr($item_html, 'li class="divider')) {
+    } elseif (stristr($item_html, 'li class="divider')) {
       $item_html = preg_replace('/<a[^>]*>.*?<\/a>/iU', '', $item_html);
-    }
-    elseif (stristr($item_html, 'li class="dropdown-header')) {
+    } elseif (stristr($item_html, 'li class="dropdown-header')) {
       $item_html = preg_replace('/<a[^>]*>(.*)<\/a>/iU', '$1', $item_html);
     }
 
@@ -50,6 +49,7 @@ class Sage_Nav_Walker extends \Walker_Nav_Menu {
 
     parent::display_element($element, $children_elements, $max_depth, $depth, $args, $output);
   }
+  // @codingStandardsIgnoreEnd
 }
 
 /**
@@ -82,10 +82,10 @@ add_filter('nav_menu_item_id', '__return_null');
  * Clean up wp_nav_menu_args
  *
  * Remove the container
- * Use Sage_Nav_Walker() by default
+ * Use SageNavWalker() by default
  */
 function nav_menu_args($args = '') {
-  $nav_menu_args = array();
+  $nav_menu_args = [];
 
   $nav_menu_args['container'] = false;