+++ templates/comments.php
@@ -1,7 +1,7 @@
 <?php
-  if (post_password_required()) {
-    return;
-  }
+if (post_password_required()) {
+  return;
+}
 ?>
 
 <section id="comments" class="comments">
@@ -9,7 +9,7 @@
     <h2><?php printf(_nx('One response to &ldquo;%2$s&rdquo;', '%1$s responses to &ldquo;%2$s&rdquo;', get_comments_number(), 'comments title', 'sage'), number_format_i18n(get_comments_number()), '<span>' . get_the_title() . '</span>'); ?></h2>
 
     <ol class="comment-list">
-      <?php wp_list_comments(array('style' => 'ol', 'short_ping' => true)); ?>
+      <?php wp_list_comments(['style' => 'ol', 'short_ping' => true]); ?>
     </ol>
 
     <?php if (get_comment_pages_count() > 1 && get_option('page_comments')) : ?>