+++ templates/content-page.php
@@ -1,2 +1,2 @@
 <?php the_content(); ?>
-<?php wp_link_pages(array('before' => '<nav class="pagination">', 'after' => '</nav>')); ?>
+<?php wp_link_pages(['before' => '<nav class="pagination">', 'after' => '</nav>']); ?>