+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 ### 8.0.0: TBD
 * Change theme name from Roots to Sage
+* Bump required PHP version to >=5.4
+* Add coding standards based on PSR-2
 * Add namespace
+* Use short array syntax
+* Use short echo syntax
 * Switch from Grunt to gulp, new workflow
 * Use wiredep for Sass and Less injection
 * Implement JSON file based asset pipeline
 * Re-organize asset file structure
+* Use the theme styles for the editor stylesheet
 * Remove theme activation, move to [wp-cli-theme-activation](https://github.com/roots/wp-cli-theme-activation)
 * Add Travis CI
 * Update to jQuery 1.11.2